#121 structuring the modules/colors
