--- modules/colors/consts/ColorErrors.ts
@@ -1,9 +1,0 @@
-const InvalidColorSpaceError = new Error("Invalid color space");
-const InvalidFormatError = new Error("Invalid format");
-const CannotConvertToHexWithAlphaError = new Error("Cannot convert to hex with alpha");
-
-export default {
-  InvalidColorSpaceError,
-  InvalidFormatError,
-  CannotConvertToHexWithAlphaError,
-};
