+++ modules/colors/data/color-errors.ts
@@ -1,0 +1,9 @@
+const InvalidColorSpaceError = new Error('Invalid color space');
+const InvalidFormatError = new Error('Invalid format');
+const CannotConvertToHexWithAlphaError = new Error('Cannot convert to hex with alpha');
+
+export default {
+    InvalidColorSpaceError,
+    InvalidFormatError,
+    CannotConvertToHexWithAlphaError,
+};
