+++ modules/colors/models/color-space.types.ts
@@ -1,0 +1,4 @@
+export interface ColorSpace {
+    id: string;
+    getNums: () => number[];
+}
